% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEasyPubPlot.R
\name{runEasyPubPlot}
\alias{runEasyPubPlot}
\title{EasyPubPlot - Easy and Publishable Ploting}
\usage{
runEasyPubPlot()
}
\description{
Interactive, customizable, and coding-free Shiny App to easily create publishable plots for scientific papers.
}
